requirements.txt
aiohttp==3.8.4
aiohttp-socks==0.8.0
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
fake-useragent==1.1.3
Faker==18.4.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.0
pluggy==1.0.0
pytest==7.2.2
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-socks==2.2.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
tomli==2.0.1
yarl==1.8.2